+++ src/ts/core/orchestration/auth/utils/loginUser.ts
@@ -19,12 +19,7 @@ export default async function loginUser(
   if (!passwortInput) throw new Error('Passwort Input nicht gefunden');
   passwort ??= passwortInput.value;
 
-  const btnLogin = document.querySelector<HTMLButtonElement>('#btnLogin');
-  if (btnLogin) btnLogin.disabled = true;
   setLoading('btnLogin');
-
-  const modalSubmitButton = modal.querySelector<HTMLButtonElement>('#btnLoginModal');
-  if (modalSubmitButton) modalSubmitButton.disabled = true;
   setLoading('btnLoginModal');
 
   const errorMessage = document.querySelector<HTMLDivElement>('#errorMessage');
@@ -32,7 +27,6 @@ export default async function loginUser(
 
   if (!navigator.onLine) {
     errorMessage.textContent = 'Keine Internetverbindung';
-    if (modalSubmitButton) modalSubmitButton.disabled = false;
     clearLoading('btnLogin', false);
     clearLoading('btnLoginModal', false);
     return;
@@ -53,7 +47,6 @@ export default async function loginUser(
       errorMessage.innerHTML = String(err);
     }
   } finally {
-    if (modalSubmitButton) modalSubmitButton.disabled = false;
     clearLoading('btnLogin', false);
     clearLoading('btnLoginModal', false);
   }